requirements.txt
streamlit
pandas
matplotlib
google-cloud-aiplatform
